Extract Span data structures to a module and add properties
