+++ Hoed.cabal
@@ -35,6 +35,7 @@ library
                        , Debug.Hoed.Fields
                        , Debug.Hoed.Prop
                        , Debug.Hoed.Serialize
+                       , Debug.Hoed.Span
                        , Text.PrettyPrint.FPretty
                        , Paths_Hoed
   build-depends:       base >= 4 && <5
@@ -46,6 +47,7 @@ library
                        , regex-tdfa
                        , directory
                        , cereal, bytestring
+                       , QuickCheck
                        , semigroups
                        , strict
                        , template-haskell
